# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Movie-Api"
    },
    "host": "localhost:8081",
    "basePath": "/movie/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "uikey": {
            "in": "header",
            "name": "uikey",
            "type": "apiKey"
        },
        "session_key": {
            "in": "header",
            "name": "apisessionkey",
            "type": "apiKey"
        }
    },
    "paths": {
...
        "/getUser": {
            "post": {
                "tags": [
                    "user"
                ],
                "description": "user(s) profile in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "profile of user in system",
                        "schema": {
                            "$ref": "#/definitions/GetUser"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/GetUser"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": [],
                        "session_key": []
                    }
                ]
            }
        },
        "/createUser": {
            "post": {
                "tags": [
                    "user"
                ],
                "description": "user(s) profile in system",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "profile of user in system",
                        "schema": {
                            "$ref": "#/definitions/CreateUser"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/CreateUser"
                        }
                    },
                    "400": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "uikey": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "GetUser": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string",
                    "uniqueItems": true
                }
            }
        },
        "CreateUser": {
            "type": "object",
            "required": [
                "userName",
                "password",
                "firstName",
                "lastName",
                "email",
                "address",
                "country"
            ],
            "properties": {
                "userName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "password": {
                    "type": "string",
                    "uniqueItems": true
                },
                "firstName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "lastName": {
                    "type": "string",
                    "uniqueItems": true
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "address": {
                    "type": "string",
                    "uniqueItems": true
                },
                "country": {
                    "type": "string",
                    "uniqueItems": true
                }
            }
        },
...
    }
}